requirements.txt
aiohttp==3.8.3
beautifulsoup4==4.11.1
Flask==2.1.2
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-Session==0.4.0
gitdb==4.0.10
GitPython==3.1.30
grequests==0.6.0
ipaddress==1.0.23
Jinja2==3.1.2
numpy==1.23.2
peewee==3.15.4
PySocks==1.7.1
Pillow
requests==2.28.2
rsa==4.9
schedule==1.1.0
selenium==4.4.3
SQLAlchemy==1.4.46
Telethon==1.27.0
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.1.2